requirements.txt
Flask==2.3.3
pdf2docx==0.5.6
pypdf==3.0.1
qrcode==7.4.2
Pillow==10.0.1
Werkzeug==2.3.7
rembg==2.0.50
google-generativeai==0.8.5
